requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
anthropic>=0.20.0
apscheduler==3.10.4
python-dotenv==1.0.0
stripe==7.8.0
httpx==0.25.1
